Put Next/PrevPageButton in a new component.. Renamed SidebarComponents.. Fixed routing problem
The Next/PrevPageButton got wrapped in a Link. With the Link, i pass an Object with pathname and a State. The state holds the ID of the genre/discovery. Importent, because it gets pushed to the next site, hence stored in history. If you go back and forth, the id of the current genre remains available.. In SidebarLinksToGenreDiscovery i hardcoded the pathname to Page=1. This ensures landing on page 1 on the clicked genre/discovery. Also added to RenderClickedLink an if-condition: if(genreId !==id){load clicked genre}. This is needed after returning from a clicked Movie. Without, the setPage(1) would kick in-> the first page of the genre would get fetched and rerendered and lastly  the wrong page will be displayed in Url.